makes "fail_hard" true by default
fail_hard is a flag that controls the behavior when getting a retryable error at submit time.
when the flag is not set retriable errors are still queued on the local instance, that will keep trying to submit it in the background.
This can be confusing for the caller of an API as normally an error