avoid holding transactions
expire old transactions awaiting to be applied
transactions were only purged when getting a validated ledger but this can take a really long time if the instance has a lot of data to download.
this was in turn causing a lot of transactions to be held and re-applied

ban transactions that fail
transactions that have no way of succeeding are marked as failed and not applied.
Bad transactions can be injected (if they were failing with a retriable error at submit time for example),
this allows to avoid keeping applying them (until the transaction gets expired frmo the local tx set).