tune down thresholds for close time consensus
With the consensus changes, the chance of some nodes considered stuck in the "ESTABLISH" phase is higher, which would cause such nodes to be very unlikely to change their position wrt close time.
This change leaves room for nodes to converge towards close time.

Trade off being made is that close time is easier to game: in the early stage of consensus in the event that the network is heavily partitioned (in the form of most nodes come up with their own close time) a group representing 30% of the nodes will influence the others. This influence would be temporary and does not impact other parts of the network that are critical (transaction set, ledger sequence number).